fix: false positive bug with github classic v1 token, fix url scanner crash, add support for boundary groups
