+++ lib/detectors/braintree/index.js
@@ -1,4 +1,4 @@
-const { surroundWithGroups } = require("../../regexPatterns");
+const { surroundWithGroups } = require("../../regexHandler");
 const Re2 = require("re2");
 
 const keywords = ["braintree"];
