+++ lib/detectors/github/classicV1.js
@@ -15,7 +15,10 @@ const scan = async (verify, data) => {
 
   for (const match of matches) {
     if (match.length !== 2) continue;
-    if (isFalsePositive(match, ["github commit"]).isFalsePositive) continue;
+    if (
+      isFalsePositive(match.input, ["commit", "github commit"]).isFalsePositive
+    )
+      continue;
 
     const resMatch = match[1].trim();
     result.rawValue = resMatch;
