+++ lib/detectors/gitlab/gitlabV1.js
@@ -1,5 +1,5 @@
 const Re2 = require("re2");
-const { surroundWithGroups } = require("../../regexPatterns");
+const { surroundWithGroups } = require("../../regexHandler");
 const { makeGitLabRequest } = require("../../analyzers/gitlab");
 
 const keywords = ["gitlab"];
