+++ lib/detectors/mixpanel/index.js
@@ -1,6 +1,6 @@
 const Re2 = require("re2");
 const axios = require("axios");
-const { surroundWithGroups } = require("../../regexPatterns");
+const { surroundWithGroups } = require("../../regexHandler");
 
 const keywords = ["mixpanel"];
 const regexGroup = surroundWithGroups(["mixpanel"]);
