+++ lib/scan.js
@@ -3,7 +3,7 @@ const fs = require("fs");
 const { scanDirectory } = require("./fileScanner");
 const { scanGitCommitsForSecrets } = require("./gitScanner");
 const { configHandler } = require("./configHandler");
-const { buildCustomDetectors } = require("./regexPatterns");
+const { buildCustomDetectors } = require("./regexHandler");
 const { analyzeRepository } = require("./urlScanner");
 const chalk = require("chalk");
