+++ lib/urlScanner.js
@@ -1,6 +1,6 @@
 const path = require("path");
 const { randomBytes } = require("crypto");
-const { execSync, spawn } = require("child_process");
+const { execSync } = require("child_process");
 const { scanDirectory } = require("./fileScanner");
 const os = require("os");
 const { scanGitCommitsForSecrets } = require("./gitScanner");
@@ -21,19 +21,6 @@ const validateAndFormatRepoUrl = (urlString) => {
   }
 };
 
-const cloneRepo = (parsedUrl, pathToSave) => {
-  return new Promise((resolve, reject) => {
-    const clone = spawn(`git clone ${parsedUrl} ${pathToSave}`);
-    clone.on("close", (code) => {
-      if (code === 0) {
-        resolve();
-      } else {
-        reject(new Error(`git clone process exited with code ${code}`));
-      }
-    });
-  });
-};
-
 const analyzeRepository = async ({
   url,
   excludedFolders,
@@ -46,7 +33,7 @@ const analyzeRepository = async ({
   try {
     // write the cloned repo to user's machine temporary directory
     const pathToSave = path.join(os.tmpdir(), filePath);
-    await cloneRepo(parsedUrl, pathToSave);
+    execSync(`git clone ${parsedUrl} ${pathToSave}`, { stdio: "pipe" });
 
     // Scan the codebase
     await scanDirectory({
