+++ lib/util.js
@@ -125,6 +125,7 @@ const isFalsePositive = (value, falsePositives) => {
   for (const fp of falsePositives) {
     if (lower.includes(fp))
       return { isFalsePositive: true, reason: `match found: ${fp}` };
+    else return { isFalsePositive: false };
   }
 };
