+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "securelog-scan",
-  "version": "1.1.7",
+  "version": "1.1.8",
   "description": "A CLI tool to scan codebases for potential secrets.",
   "main": "bin/sls.js",
   "author": {
